+++ packages/bridge-controller/src/bridge-controller.test.ts
@@ -41,6 +41,7 @@ import mockBridgeQuotesErc20Native from '../tests/mock-quotes-erc20-native.json'
 import mockBridgeQuotesNativeErc20Eth from '../tests/mock-quotes-native-erc20-eth.json';
 import mockBridgeQuotesNativeErc20 from '../tests/mock-quotes-native-erc20.json';
 import mockBridgeQuotesSolErc20 from '../tests/mock-quotes-sol-erc20.json';
+import { advanceToNthTimerThenFlush } from '../tests/mock-sse';
 
 const EMPTY_INIT_STATE = DEFAULT_BRIDGE_CONTROLLER_STATE;
 
@@ -391,19 +392,28 @@ describe('BridgeController', function () {
     const hasSufficientBalanceSpy = jest
       .spyOn(balanceUtils, 'hasSufficientBalance')
       .mockResolvedValue(true);
-    messengerMock.call.mockReturnValue({
-      address: '0x123',
-      provider: jest.fn(),
-      currencyRates: {},
-      marketData: {},
-      conversionRates: {},
-      remoteFeatureFlags: {
-        bridgeConfig: {
-          ...bridgeConfig,
-          sse: { enabled: true, minimumVersion: '13.9.0' },
-        },
+    messengerMock.call.mockImplementation(
+      (...args: Parameters<BridgeControllerMessenger['call']>) => {
+        switch (args[0]) {
+          case 'AuthenticationController:getBearerToken':
+            return 'AUTH_TOKEN';
+          default:
+            return {
+              address: '0x123',
+              provider: jest.fn(),
+              currencyRates: {},
+              marketData: {},
+              conversionRates: {},
+              remoteFeatureFlags: {
+                bridgeConfig: {
+                  ...bridgeConfig,
+                  sse: { enabled: true, minimumVersion: '13.9.0' },
+                },
+              },
+            } as never;
+        }
       },
-    } as never);
+    );
 
     const fetchBridgeQuotesSpy = jest
       .spyOn(fetchUtils, 'fetchBridgeQuotes')
@@ -511,6 +521,7 @@ describe('BridgeController', function () {
       },
       expect.any(AbortSignal),
       BridgeClientId.EXTENSION,
+      'AUTH_TOKEN',
       mockFetchFn,
       BRIDGE_PROD_API_BASE_URL,
       null,
@@ -974,14 +985,23 @@ describe('BridgeController', function () {
     const hasSufficientBalanceSpy = jest
       .spyOn(balanceUtils, 'hasSufficientBalance')
       .mockResolvedValue(false);
-    messengerMock.call.mockReturnValue({
-      address: '0x123',
-      provider: jest.fn(),
-      currentCurrency: 'usd',
-      currencyRates: {},
-      marketData: {},
-      conversionRates: {},
-    } as never);
+    messengerMock.call.mockImplementation(
+      (...args: Parameters<BridgeControllerMessenger['call']>) => {
+        switch (args[0]) {
+          case 'AuthenticationController:getBearerToken':
+            return 'AUTH_TOKEN';
+          default:
+            return {
+              address: '0x123',
+              provider: jest.fn(),
+              currentCurrency: 'usd',
+              currencyRates: {},
+              marketData: {},
+              conversionRates: {},
+            } as never;
+        }
+      },
+    );
     jest
       .spyOn(selectors, 'selectIsAssetExchangeRateInState')
       .mockReturnValue(true);
@@ -1065,6 +1085,7 @@ describe('BridgeController', function () {
       },
       expect.any(AbortSignal),
       BridgeClientId.EXTENSION,
+      'AUTH_TOKEN',
       mockFetchFn,
       BRIDGE_PROD_API_BASE_URL,
       null,
@@ -1360,6 +1381,124 @@ describe('BridgeController', function () {
     );
   });
 
+  it('updateBridgeQuoteRequestParams should include undefined Authentication header if getBearerToken throws an error', async function () {
+    jest.useFakeTimers();
+    const startPollingSpy = jest.spyOn(bridgeController, 'startPolling');
+    messengerMock.call.mockImplementation(
+      (...args: Parameters<BridgeControllerMessenger['call']>) => {
+        switch (args[0]) {
+          case 'AuthenticationController:getBearerToken':
+            throw new Error(
+              'AuthenticationController:getBearerToken not implemented',
+            );
+          default:
+            return {
+              address: '0x123',
+              provider: jest.fn(),
+              currentCurrency: 'usd',
+              currencyRates: {},
+              marketData: {},
+              conversionRates: {},
+            } as never;
+        }
+      },
+    );
+    jest
+      .spyOn(selectors, 'selectIsAssetExchangeRateInState')
+      .mockReturnValue(true);
+
+    const fetchBridgeQuotesSpy = jest
+      .spyOn(fetchUtils, 'fetchBridgeQuotes')
+      .mockImplementationOnce(async () => {
+        return await new Promise((resolve) => {
+          return setTimeout(() => {
+            resolve({
+              quotes: mockBridgeQuotesNativeErc20Eth as never,
+              validationFailures: [],
+            });
+          }, 5000);
+        });
+      });
+
+    const quoteParams = {
+      srcChainId: '0x1',
+      destChainId: '0xa',
+      srcTokenAddress: '0x0000000000000000000000000000000000000000',
+      destTokenAddress: '0x123',
+      srcTokenAmount: '1000000000000000000',
+      walletAddress: '0x123',
+      slippage: 0.5,
+    };
+
+    await bridgeController.updateBridgeQuoteRequestParams(
+      quoteParams,
+      metricsContext,
+    );
+
+    await advanceToNthTimerThenFlush();
+
+    expect(startPollingSpy).toHaveBeenCalledTimes(1);
+    expect(fetchBridgeQuotesSpy.mock.calls[0][3]).toBeUndefined();
+  });
+
+  it('updateBridgeQuoteRequestParams should include auth token as Authentication header', async function () {
+    jest.useFakeTimers();
+    const startPollingSpy = jest.spyOn(bridgeController, 'startPolling');
+    messengerMock.call.mockImplementation(
+      (...args: Parameters<BridgeControllerMessenger['call']>) => {
+        switch (args[0]) {
+          case 'AuthenticationController:getBearerToken':
+            return 'AUTH_TOKEN';
+          default:
+            return {
+              address: '0x123',
+              provider: jest.fn(),
+              currentCurrency: 'usd',
+              currencyRates: {},
+              marketData: {},
+              conversionRates: {},
+            } as never;
+        }
+      },
+    );
+    jest
+      .spyOn(selectors, 'selectIsAssetExchangeRateInState')
+      .mockReturnValue(true);
+
+    const fetchBridgeQuotesSpy = jest
+      .spyOn(fetchUtils, 'fetchBridgeQuotes')
+      .mockImplementationOnce(async () => {
+        return await new Promise((resolve) => {
+          return setTimeout(() => {
+            resolve({
+              quotes: mockBridgeQuotesNativeErc20Eth as never,
+              validationFailures: [],
+            });
+          }, 5000);
+        });
+      });
+
+    const quoteParams = {
+      srcChainId: '0x1',
+      destChainId: '0xa',
+      srcTokenAddress: '0x0000000000000000000000000000000000000000',
+      destTokenAddress: '0x123',
+      srcTokenAmount: '1000000000000000000',
+      walletAddress: '0x123',
+      slippage: 0.5,
+    };
+
+    await bridgeController.updateBridgeQuoteRequestParams(
+      quoteParams,
+      metricsContext,
+    );
+
+    await advanceToNthTimerThenFlush();
+
+    expect(startPollingSpy).toHaveBeenCalledTimes(1);
+    expect(fetchBridgeQuotesSpy.mock.calls[0][3]).toBe('AUTH_TOKEN');
+  });
+
   it.each([
     [
       'should append l1GasFees if srcChain is 10 and srcToken is erc20',
@@ -1411,10 +1550,19 @@ describe('BridgeController', function () {
       const hasSufficientBalanceSpy = jest
         .spyOn(balanceUtils, 'hasSufficientBalance')
         .mockResolvedValue(false);
-      messengerMock.call.mockReturnValue({
-        address: '0x123',
-        provider: jest.fn(),
-      } as never);
+      messengerMock.call.mockImplementation(
+        (...args: Parameters<BridgeControllerMessenger['call']>) => {
+          switch (args[0]) {
+            case 'AuthenticationController:getBearerToken':
+              return 'AUTH_TOKEN';
+            default:
+              return {
+                address: '0x123',
+                provider: jest.fn(),
+              } as never;
+          }
+        },
+      );
 
       for (const [index, quote] of quoteResponse.entries()) {
         if (tradeL1GasFeeError && index === 0) {
@@ -1500,6 +1648,7 @@ describe('BridgeController', function () {
         },
         expect.any(AbortSignal),
         BridgeClientId.EXTENSION,
+        'AUTH_TOKEN',
         mockFetchFn,
         BRIDGE_PROD_API_BASE_URL,
         null,
@@ -1799,6 +1948,10 @@ describe('BridgeController', function () {
         ): ReturnType<BridgeControllerMessenger['call']> => {
           const [actionType, params] = args;
 
+          if (actionType === 'AuthenticationController:getBearerToken') {
+            return 'AUTH_TOKEN';
+          }
+
           if (actionType === 'AccountsController:getAccountByAddress') {
             if (isEvmAccount) {
               return {
@@ -1919,6 +2072,13 @@ describe('BridgeController', function () {
       // Loading state
       jest.advanceTimersByTime(201);
       await flushPromises();
+
+      // Wait for JWT token retrieval
+      if (!isEvmAccount) {
+        jest.advanceTimersToNextTimer();
+        await flushPromises();
+      }
+
       expect(bridgeController.state).toStrictEqual(
         expect.objectContaining({
           quotesLoadingStatus: RequestStatus.LOADING,
@@ -2597,6 +2757,7 @@ describe('BridgeController', function () {
         UnifiedSwapBridgeEventName.StatusValidationFailed,
         {
           failures: ['Failed to submit tx'],
+          refresh_count: 0,
         },
       );
       expect(trackMetaMetricsFn).toHaveBeenCalledTimes(1);
@@ -2643,6 +2804,7 @@ describe('BridgeController', function () {
           } as never;
         },
       );
+      bridgeController.setLocation(MetaMetricsSwapsEventSource.TrendingExplore);
     });
 
     it('should not track the event if the account keyring type is not set', async () => {
@@ -2741,6 +2903,7 @@ describe('BridgeController', function () {
             },
           },
         });
+      (messengerMock.call as jest.Mock).mockResolvedValueOnce('AUTH_TOKEN');
       (messengerMock.call as jest.Mock).mockReturnValueOnce(() => ({
         address: '0x123',
       }));
@@ -2801,6 +2964,7 @@ describe('BridgeController', function () {
             },
             null,
             "extension",
+            "AUTH_TOKEN",
             [Function],
             "https://bridge.api.cx.metamask.io",
             "perps",
@@ -2897,6 +3061,7 @@ describe('BridgeController', function () {
             },
             null,
             "extension",
+            "AUTH_TOKEN",
             [Function],
             "https://bridge.api.cx.metamask.io",
             "perps",
@@ -2950,6 +3115,7 @@ describe('BridgeController', function () {
             },
             null,
             "extension",
+            "AUTH_TOKEN",
             [Function],
             "https://bridge.api.cx.metamask.io",
             null,