+++ packages/assets-controllers/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Uncategorized
+
+- chore(assets-controllers): replace `Sinon` with `Jest` fake timers and mocks ([#7981](https://github.com/MetaMask/core/pull/7981))
+
 ### Changed
 
 - Bump `@metamask/phishing-controller` from `^16.2.0` to `^16.3.0` ([#7979](https://github.com/MetaMask/core/pull/7979))